Upload ex12.c
This program starts from the given path folder and run in all subfolder
and give grades to c programs that found inside. Only if the compilation
ok and output correct we will get 100. After execution all temporary
files will be removed and grades will be written to 'results.csv' file.